python_server/requirements.txt
aniso8601==10.0.1
blinker==1.9.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.6
Flask==3.1.1
flask-cors==6.0.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.4.0
packaging==25.0
pluggy==1.6.0
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.1
pytest==8.4.1
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.43
typing_extensions==4.14.1
Werkzeug==3.1.3
